requirements.txt
pyspark>=3.5.0
pandas>=2.2.0
pydeck>=0.9.1
streamlit>=1.31.0
PyYAML>=6.0
pyarrow>=14.0.0
pytest>=7.4.0
